requirements.txt
asgiref==3.11.1
Django==6.0.4
djangorestframework==3.17.1
numpy==2.4.4
openai
pandas==3.0.2
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.5
